✨ feat: Icon 컴포넌트 이동 (components-deprecated -> components)
