+++ src/components-deprecated/reusable/ChatInput/TodoListInput/TodoItemInput/index.tsx
@@ -1,7 +1,6 @@
 import styled from '@emotion/styled';
 
-import Icon from '@src/components-deprecated/reusable/Icon';
-
+import { Icon } from '@src/components-deprecated/reusable';
 const TodoItemInput = () => {
   return (
     <Wrapper>
