+++ src/components-deprecated/reusable/index.ts
@@ -4,7 +4,7 @@ export { default as Button } from './Button';
 // 추후 삭제될 코드이므로 임시로 처리
 // eslint-disable-next-line import/no-cycle
 export { default as ChatInput } from './ChatInput';
-export { default as Icon } from './Icon';
+export { default as Icon } from '../../components/reusable/Icon';
 export { default as Modal } from './Modal';
 export { default as ModalLayout } from './ModalLayout';
