+++ src/components/reusable/Icon/index.ts
@@ -1,0 +1,2 @@
+export { default } from './Icon';
+export type { IconNameType, SvgIconProps } from './Icon.types';
